feat(query): Improve explain test compatibility from 57% to ~65%
Fix execute explain metrics (planExecutions/sizeOfResult), top-level
aggregate decomposition, average node chain explain, inline array
aggregate sources, and execute metrics u64 casting across all plan nodes.

- Fix planExecutions to count top-level query operations (N+1) matching
  Go's operationNode Next() semantics
- Fix sizeOfResult to count number of query results, not documents
- Decompose top-level averageNode into sumNode + countNode + averageNode
- Wrap non-top-level averageNode explain in countNode → sumNode chain
- Add SumSourceMeta, MaxSourceMeta, MinSourceMeta for inline array agg
- Cast all explain metrics to u64 for Go compatibility
- Handle execute vs default explain for top-level aggregate nodes
- Strip aggregate iterations when unwrapping for topLevelNode
- Add stub FFI functions for p2p_add/remove/get_all_documents and
  truncate_collection
- Handle _commits system collection in execute explain path

+++ crates/query/src/plan/count.rs
@@ -295,13 +295,11 @@ impl PlanNode for CountNode {
     fn explain_execute_inner(&self) -> JsonValue {
         let mut obj = serde_json::Map::new();
 
-        // Go DefraDB execute format: iterations
         obj.insert(
             "iterations".to_string(),
-            serde_json::json!(self.exec_info.iterations),
+            serde_json::json!(self.exec_info.iterations as u64),
         );
 
-        // Recursively explain child node with execution info
         let child_explain = self.source.explain_execute();
         if let Some(child_obj) = child_explain.as_object() {
             for (key, value) in child_obj {
